refactor: split oversized cleanup targets
