+++ admin-ui/src/features/debug/index.ts
@@ -1,0 +1,2 @@
+export { DebugPanel } from './DebugPanel';
+export type { DebugPanelProps } from './DebugPanel';
