+++ admin-ui/src/features/governance/index.ts
@@ -1,0 +1,2 @@
+export { GovernancePanel } from './GovernancePanel';
+export type { GovernancePanelProps, GovernanceSummary } from './GovernancePanel';
