+++ admin-ui/src/features/instances/index.ts
@@ -1,0 +1,2 @@
+export { InstancesPanel } from './InstancesPanel';
+export type { InstanceUpdateNotice, InstancesPanelProps } from './InstancesPanel';
