+++ admin-ui/src/features/openapi/index.ts
@@ -1,0 +1,2 @@
+export { OpenApiPanel } from './OpenApiPanel';
+export type { OpenApiPanelProps } from './OpenApiPanel';
