+++ scripts/ci/check_rust_test_file_lines.py
@@ -11,11 +11,7 @@ MAX_TEST_LINES = 2000
 
 # Existing oversized Rust test modules. These are allowed to stay at or below
 # the recorded baseline, but any growth must be split into focused test files.
-LEGACY_BASELINES = {
-    "crates/dcc-mcp-gateway/src/gateway/admin/tests.rs": 4883,
-    "crates/dcc-mcp-cli/tests/cli_e2e.rs": 3342,
-    "crates/dcc-mcp-gateway/src/gateway/handlers/rest_impl_tests.rs": 2141,
-}
+LEGACY_BASELINES = {}
 
 
 def is_rust_test_file(path: Path) -> bool:
